Expose HMAC instead of keyed BLAKE2
The HMAC construction is well defined, and is available for every hash
function. While BLAKE2 supports a keyed version which is another MAC,
the `hmac_` functions should compute HMAC.

This changes the BLAKE hmac functions so that they return the proper
values (this is a breaking change obviously), and add new `Keyed`
submodules.

It is possible to cross-check against `openssl` by using the following
commands:

    echo -n 'Hi There' | openssl dgst -blake2b512 -hmac 'Salut'
    echo -n 'Hi There' | openssl dgst -blake2s256 -hmac 'Salut'

Closes #46