Delete the old MirageOS 3 trick
Such trick is to compile digestif with ocaml-solo5 and distribute
it for MirageOS 3 then. We need to patch the generated META file
from `dune` to finally install something good for MirageOS 3.

It seems clear that at this date, we don't really need to continue
the support of MirageOS. It seems that such trick bothers dune
cache and installation on some platform. Let's just remove that.